Add Briefly.cleanup
It is sometimes useful to cleanup temporary files created in the context
of the current process without waiting for it to shutdown.

For example, one may have a long-lived worker process that needs to
regularly remove all files created via Briefly to avoid using all disk
space over time.

Although this could be achieved by keeping track of those files in the
worker process directly, it is done much more cleanly by leveraging the
already existing Briefly ETS table.

The function returns the list of paths tracked by Briefly for the
current process in case the client wants to perform further processing.