Avoid folder naming conflicts with monotonic time and random padding
Switches to using monotonic time and random padding from crypto
to ensure that folders are created uniquely even if created in same moment by
by different users or different instances of server.

By adding 20 chars of random padding along with monotonic time, the chance of
collisions is so small as to be negligible.

Further work could add a safety check that it retries folder creation if first one
fails.

Fixes: https://github.com/CargoSense/briefly/issues/16