Add an option for returning a device similar to mkstemp
If the temporary directory is world writable, in order to securely open
a temporary file, you need to create it and open it at the same time and
have exclusive access to it. It remains secure as long as you hold onto
the file descriptor returned by open. Requesting a device returns the io
device and the path, rather than just the path.

Note that the io device is owned by the pid that called open(), and will
be closed when that process exits. There is no way to hand off the
device to another pid. If a temporary device needs to be long lived, it
needs to be opened by a process that is at least as long lived.